# perf(@formatjs/ecma402-abstract): cache NumberFormat instances for 20x speed-up when formatting dates