Fixes #2807 
Fixes #2813 

It appears that creating `new Intl.NumberFormat` instances is extremely slow. Caching them by locale key seems safe, and this dramatically increases performance.

Before:
```
format (polyfill) x 15,193 ops/sec ±2.86% (87 runs sampled)
format (native) x 551,023 ops/sec ±4.01% (83 runs sampled)
```

After:
```
format (polyfill) x 326,105 ops/sec ±0.42% (98 runs sampled)
format (native) x 643,658 ops/sec ±0.62% (93 runs sampled)
```

cc @longlho