@longlho 2021-04-13
This should be resolved with caching the constructor upstream right?