@andreialecu 2021-04-13
Not sure what you mean by caching the constructor upstream. 

I'm using the benchmark that is part of the repo, which caches the `DateTimeFormat` constructor.