@longlho 2021-04-13
As in user should cache new Intl.NumberFormat regardless of if it's polyfills or native