@andreialecu 2021-04-13
I'm still confused by what you mean. This isn't about caching `Intl.NumberFormat`, the method I modified creates new instances of `Intl.NumberFormat` as part of formatting a date.

Feel free to review it when time permits.