@andreialecu 2021-04-13
Just an opinion, but I think this repository and everyone using and contributing to it would greatly benefit if the handling of issues was taken more seriously.

I'd have appreciated not spending so much time in #2807 and #2813 testing and replying to one-liners just to re-explain something that could be understood just by treating things seriously. Thankfully the issue was finally acknowledged and some steps started to be taken in: https://github.com/formatjs/formatjs/commit/812bcb6bd202ccd2dd9222347c09fa91ca2073a5 

#2807 in particular was closed and locked with a comment in which the first paragraph was proven incorrect by: https://github.com/formatjs/formatjs/commit/463420c21ca9c64f629f31fee2dc63031be9122c, while other comments like https://github.com/facebook/hermes/issues/23#issuecomment-815992481 are simply not backed by any real numbers.

I even read the spec, and don't see how this could cause any problems, considering the changes simply add a private cache that is only used in one specific place.

My final hope was that, after kindly asking, you'd spend some time to review a 20x performance increase that actually unblocks a lot of people and stops wasting a lot of resources. I also noticed a dramatic reduction in GCs by using this patch.

I'll deal with using a patch via [patch-package](https://www.npmjs.com/package/patch-package), but this would've been useful to the people in https://github.com/facebook/hermes/issues/23 as well.