@longlho 2021-04-13
I'd suggest discussing your direction before spending time on implementation. As I mentioned we try to keep it as close to the spec as possible and slipping in 463420c21ca9c64f629f31fee2dc63031be9122c is already breaking that but it's small enough I'm ok with it. Introducing new caching is a much bigger change:
1. The behavior is not defined in the spec.
2. It creates a side effect which is a global cache with no ability to flush.
3. Underlying data can change, it's not guaranteed to be static by the cache. Even in this polyfill you can add/invalidate data at runtime.

Basically my rule of thumb is, if you're changing the polyfill, please link to the line either in the spec, or in reference implementations such as `icu4j`/`icu4c`.