@andreialecu 2021-04-13
The time spent implementing was useful to me. This patch is actually something I need and is the only way to even make this feasible to use. So I'm good on that.

The only time I feel is wasted is the time I spent trying to communicate about this issue.

Regarding caching, from what I can see in the spec https://402.ecma-international.org/7.0/#sec-partitiondatetimepattern:

```
Let locale be dateTimeFormat.[[Locale]].
Let nfOptions be ObjectCreate(null).
Perform ! CreateDataPropertyOrThrow(nfOptions, "useGrouping", false).
Let nf be ? Construct(%NumberFormat%, « locale, nfOptions »).
Let nf2Options be ObjectCreate(null).
Perform ! CreateDataPropertyOrThrow(nf2Options, "minimumIntegerDigits", 2).
Perform ! CreateDataPropertyOrThrow(nf2Options, "useGrouping", false).
Let nf2 be ? Construct(%NumberFormat%, « locale, nf2Options »).
```

The `nfOptions` to construct `nf` and `nf2` are static and they never change. The only thing that varies is the `locale`.

If users add or remove locale data at runtime, then that can be a problem, but that seems to go against the spec as well. I wasn't able to find anything in the spec that mentions this being a supported scenario.