@longlho 2021-04-13
> If users add or remove locale data at runtime, then that can be a problem, but that seems to go against the spec as well.

The spec is intentionally being vague about data source/shape as we don't want to tie it to a specific repo like CLDR.